AI + Lighting
Slight tad more AI work. Made a class for melee enemy to take damage etc, likely going to be generic soon and shared.
Started working on prettying up the scene as I'm fed up of stock Synty asset scene looking bad and I want to make some promo videos.
Didn't push LightingData, want to see if it's essential or if it's fine if I just rebuild Lighting after cloning.

+++ Assets/Scripts/Characters/ICharacter.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface ICharacter
 {
-    public void TakeDamage(float damage);
-    public void Heal(float heal);
+    public void TakeDamage(int damage);
+    public void Heal(int heal);
 }
